Create greedy.c
This code is really close to working correctly. Some inputs work fine (3.43, 10.01, .18), and output the correct results, while others like 10.00, or 20.00, or .15, will not compute the correct amount of coins needed. I cannot find the correlation behind the incorrect outputs. 

I'd guess it has something to do with my float and int designations...